+++ src/store/offers/selectors/selectors.js
@@ -45,3 +45,9 @@ export const getIsNeedShowError = createSelector(
     (status, offers, citiOffers) => status === LoadStatus.FAILURE
       || offers.length === 0
       || citiOffers === 0);
+
+export const getIsNeedRedirect = createSelector(
+  [getStatus, getOffers, getOffersByCity],
+  (status, offers, citiOffers) => status !== LoadStatus.FAILURE
+    || offers.length > 0
+    || citiOffers > 0);
